[![GEOS-11089](https://badgen.net/badge/JIRA/GEOS-11089/0052CC)](https://osgeo-org.atlassian.net/browse/GEOS-11089)<!-- PR-BADGE: PLEASE DO NOT REMOVE THIS COMMENT -->

Use a specialization of `CatalogInfoLookup` for `NamespaceInfo` that speeds up lookup of namespaces by URI.

With this patch and the one for [GEOS-11088](https://github.com/geoserver/geoserver/pull/7050), starting up from a catalog with 120k workspaces and 600k layers, goes from 18 to 4 minutes with the default geoserver loader, and down to less than a minute with the [datadir-catalog-loader](https://github.com/geoserver/geoserver/tree/580221af6da9aedcd4e0333ea205e92277ca4c44/src/community/datadir-catalog-loader) community module.

```
                                     before   after
default loader                        1113s    294s
datadir-catalog-loader (4 trheads)     587s     98s
datadir-catalog-loader (8 trheads)     334s     69s
datadir-catalog-loader (16 trheads)    245s     57s
```

<!--Include a few sentences describing the overall goals for this Pull Request-->
  
<!-- Please help our volunteers reviewing this PR by completing the following items. 
Ask in a comment if you have troubles with any of them. -->

# Checklist

- [x] I have read the [contribution guidelines](https://github.com/geoserver/geoserver/blob/main/CONTRIBUTING.md).
- [x] I have sent a [Contribution Licence Agreement](https://docs.geoserver.org/latest/en/developer/policies/committing.html) (not required for small changes, e.g., fixing typos in documentation).
- [x] First PR targets the `main` branch (backports managed later; ignore for branch specific issues).
- [x] All the build checks are green ([see automated QA checks](https://docs.geoserver.org/latest/en/developer/qa-guide/index.html)).

For core and extension modules:

- [x] New unit tests have been added covering the changes.
- [ ] [Documentation](https://github.com/geoserver/geoserver/tree/main/doc/en/user/source) has been updated (if change is visible to end users).
- [ ] The [REST API docs](https://github.com/geoserver/geoserver/tree/main/doc/en/api/1.0.0) have been updated (when changing configuration objects or the REST controllers).
- [x] There is an issue in the [GeoServer Jira](https://osgeo-org.atlassian.net/browse/GEOS/summary) (except for changes that do not affect administrators or end users in any way).
- [x] Commit message(s) must be in the form ``[GEOS-XYZWV] Title of the Jira ticket``.
- [x] Bug fixes and small new features are presented as a single commit.
- [x] Each commit has a single objective (if there are multiple commits, each has a separate JIRA ticket describing its goal).

<!--Submitting the PR does not require you to check all items, but by the time it gets merged, they should be either satisfied or not applicable.-->