@aaime 2023-08-07
Looking good. So this helps in case there is a ton of workspaces, which may be the case for a SAAS setup (someone selling thin  slices of a GeoServer installation, each to a customer with a small amount of layers).